# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "توفير API",
  "description": "توفير — منصة يمنية للخصومات وطلب الوجبات. منصة متعددة الأدوار: العميل (تصفّح المنتجات اليمنية وطلب الوجبات والاشتراك بعضوية الخصم 30%)، المالك (إدارة منشأته ومنتجاته مع استيراد Excel)، والمشرف (إدارة النظام بالكامل وسجل التدقيق). الاشتراك بالعضوية يتم عبر رفع صورة التحويل بانتظار موافقة يدوية من المشرف.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/facilities": {
   "get": {
    "tags": [
     "customer-facilities"
    ],
    "summary": "Get Facilities",
    "description": "المنشآت العامة الموافق عليها والظاهرة، مع ترتيب اختياري حسب الأقرب مسافة.",
    "operationId": "get_facilities_api_v1_facilities_get",
    "parameters": [
     {
      "name": "region_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "فلترة حسب المنطقة (اختياري)",
       "title": "Region Id"
      },
      "description": "فلترة حسب المنطقة (اختياري)"
     },
     {
      "name": "lat",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "description": "خط العرض للترتيب بالأقرب",
       "title": "Lat"
      },
      "description": "خط العرض للترتيب بالأقرب"
     },
     {
      "name": "lng",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "description": "خط الطول للترتيب بالأقرب",
       "title": "Lng"
      },
      "description": "خط الطول للترتيب بالأقرب"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/FacilityOut"
         },
         "title": "Response Get Facilities Api V1 Facilities Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/login": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Admin Login",
    "operationId": "admin_login_api_v1_admin_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AdminLogin"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/owner/register": {
   "post": {
    "tags": [
     "owner-auth"
    ],
    "summary": "Owner Register Endpoint",
    "description": "Owner self-registration — account active, facility pending approval.",
    "operationId": "owner_register_endpoint_api_v1_owner_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OwnerRegister"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OwnerRegisterOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/owner/login": {
   "post": {
    "tags": [
     "owner-auth"
    ],
    "summary": "Owner Login Endpoint",
    "operationId": "owner_login_endpoint_api_v1_owner_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OwnerLogin"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AdminLogin": {
    "properties": {
     "identifier": {
      "type": "string",
      "minLength": 1,
      "title": "Identifier"
     },
     "password": {
      "type": "string",
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "identifier",
     "password"
    ],
    "title": "AdminLogin",
    "description": "``identifier`` accepts either the admin email or phone number."
   },
...
   "CardBrief": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "discount_rate": {
      "type": "number",
      "title": "Discount Rate"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "discount_rate"
    ],
    "title": "CardBrief"
   },
...
   "FacilityOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "type": {
      "$ref": "#/components/schemas/FacilityType"
     },
     "region_id": {
      "type": "integer",
      "title": "Region Id"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "is_visible": {
      "type": "boolean",
      "title": "Is Visible"
     },
     "display_order": {
      "type": "integer",
      "title": "Display Order"
     },
     "cards": {
      "items": {
       "$ref": "#/components/schemas/CardBrief"
      },
      "type": "array",
      "title": "Cards",
      "default": []
     },
     "owner_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Owner Id"
     },
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "latitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latitude"
     },
     "longitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Longitude"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "working_hours": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Working Hours"
     },
     "image_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Image Url"
     },
     "is_approved": {
      "type": "boolean",
      "title": "Is Approved",
      "default": false
     },
     "rejection_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rejection Reason"
     },
     "approved_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Approved At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "type",
     "region_id",
     "is_visible",
     "display_order",
     "created_at"
    ],
    "title": "FacilityOut"
   },
...
   "FacilityType": {
    "type": "string",
    "enum": [
     "restaurant",
     "cafe"
    ],
    "title": "FacilityType"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "OwnerLogin": {
    "properties": {
     "identifier": {
      "type": "string",
      "minLength": 1,
      "title": "Identifier"
     },
     "password": {
      "type": "string",
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "identifier",
     "password"
    ],
    "title": "OwnerLogin",
    "description": "``identifier`` accepts either the owner email or phone number."
   },
   "OwnerRegister": {
    "properties": {
     "full_name": {
      "type": "string",
      "maxLength": 160,
      "minLength": 2,
      "title": "Full Name"
     },
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "phone": {
      "type": "string",
      "maxLength": 40,
      "title": "Phone"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 6,
      "title": "Password"
     },
     "password_confirm": {
      "type": "string",
      "maxLength": 128,
      "minLength": 6,
      "title": "Password Confirm"
     },
     "facility_name": {
      "type": "string",
      "maxLength": 160,
      "minLength": 2,
      "title": "Facility Name"
     },
     "facility_type": {
      "$ref": "#/components/schemas/FacilityType"
     },
     "region_id": {
      "type": "integer",
      "title": "Region Id"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "address": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "phone_facility": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 40
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone Facility"
     },
     "working_hours": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Working Hours"
     },
     "image_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Image Url"
     }
    },
    "type": "object",
    "required": [
     "full_name",
     "email",
     "phone",
     "password",
     "password_confirm",
     "facility_name",
     "facility_type",
     "region_id"
    ],
    "title": "OwnerRegister",
    "description": "Owner self-registration with facility data.\n\nThe owner account is active immediately (can login) but the facility\nis created with ``is_approved=False`` pending admin review."
   },
   "OwnerRegisterOut": {
    "properties": {
     "detail": {
      "type": "string",
      "title": "Detail"
     },
     "status_code": {
      "type": "integer",
      "title": "Status Code",
      "default": 201
     },
     "user_id": {
      "type": "integer",
      "title": "User Id"
     },
     "facility_id": {
      "type": "integer",
      "title": "Facility Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "detail",
     "user_id",
     "facility_id",
     "status"
    ],
    "title": "OwnerRegisterOut",
    "description": "Owner registration response — no membership, includes facility_id and status."
   },
...
   "TokenOut": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     }
    },
    "type": "object",
    "required": [
     "access_token"
    ],
    "title": "TokenOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}